Finish renaming and removal to replace old chip
