+++ extensions/rv32im/circuit/src/lib.rs
@@ -5,7 +5,7 @@ mod base_alu;
 mod branch_eq;
 mod branch_lt;
 mod divrem;
-mod hintstore_new;
+mod hintstore;
 mod jal_lui;
 mod jalr;
 mod less_than;
@@ -20,7 +20,7 @@ pub use base_alu::*;
 pub use branch_eq::*;
 pub use branch_lt::*;
 pub use divrem::*;
-pub use hintstore_new::*;
+pub use hintstore::*;
 pub use jal_lui::*;
 pub use jalr::*;
 pub use less_than::*;
